+++ src/str.h
@@ -11,15 +11,15 @@ typedef struct _string_ {
 # define STR_CONST	((uindex) ((UINDEX_MAX + 1L) >> 1))
 # define STR_REF	(UINDEX_MAX - STR_CONST)
 
-extern void	str_init	P((void));
-extern string  *str_new		P((char*, long));
+extern void		str_init	P((void));
+extern string	       *str_new		P((char*, long));
 # define str_ref(s)	((s)->ref++)
-extern void	str_del		P((string*));
+extern void		str_del		P((string*));
 
-extern long	str_put		P((char*, long));
-extern void	str_clear	P((void));
+extern long		str_put		P((char*, long));
+extern void		str_clear	P((void));
 
-extern int	str_cmp		P((string*, string*));
-extern string  *str_add		P((string*, string*));
-extern int	str_index	P((string*, Int));
-extern string  *str_range	P((string*, Int, Int));
+extern int		str_cmp		P((string*, string*));
+extern string	       *str_add		P((string*, string*));
+extern unsigned short	str_index	P((string*, Int));
+extern string	       *str_range	P((string*, Int, Int));
